# custos/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse:
    properties:
      code:
        type: string
      fields:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
...
  github_com_julesChu12_fly_custos_internal_application_dto.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  github_com_julesChu12_fly_custos_internal_application_dto.RefreshRequest:
    properties:
      refresh_token:
        type: string
      session_id:
        type: string
    required:
    - refresh_token
    - session_id
    type: object
  github_com_julesChu12_fly_custos_internal_application_dto.RegisterRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  github_com_julesChu12_fly_custos_internal_application_dto.SuccessResponse:
    properties:
      data: {}
    type: object
...
  internal_interface_http_handler_auth.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  internal_interface_http_handler_auth.ValidatePasswordRequest:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  internal_interface_http_handler_auth.ValidatePasswordResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      is_valid:
        type: boolean
      score:
        type: integer
      strength_message:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: support@fly.com
    name: API Support
    url: http://www.fly.com/support
  description: 用户认证和授权服务API文档，提供用户注册、登录、OAuth、RBAC等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Custos User & Authentication Service API
  version: "1.0"
paths:
...
  /api/v1/admin/users/{user_id}/password/check:
    post:
      consumes:
      - application/json
      description: Checks the password strength for a specific user (admin access
        required)
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Password to check
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interface_http_handler_auth.ValidatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interface_http_handler_auth.ValidatePasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Check password strength for user
      tags:
      - password
  /api/v1/password/change:
    post:
      consumes:
      - application/json
      description: Changes the authenticated user's password
      parameters:
      - description: Password change details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interface_http_handler_auth.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - password
...
  /api/v1/password/validate:
    post:
      consumes:
      - application/json
      description: Validates a password against the current security policy
      parameters:
      - description: Password to validate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interface_http_handler_auth.ValidatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interface_http_handler_auth.ValidatePasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      summary: Validate password strength
      tags:
      - password
  /auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名/邮箱和密码登录
      parameters:
      - description: 登录凭证
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      summary: 用户登录
      tags:
      - 认证
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用刷新令牌获取新的访问令牌
      parameters:
      - description: 刷新令牌
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      summary: 刷新访问令牌
      tags:
      - 认证
  /auth/register:
    post:
      consumes:
      - application/json
      description: 创建新用户账户
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      summary: 用户注册
      tags:
      - 认证
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...